+++ src/PVOutput.Net/Requests/Modules/AddOutputRequest.cs
@@ -26,7 +26,7 @@ namespace PVOutput.Net.Requests.Modules
             ["cd"] = FormatHelper.GetEnumerationDescription(Output.Condition),
             ["tm"] = FormatHelper.GetValueAsString(Output.MinimumTemperature),
             ["tx"] = FormatHelper.GetValueAsString(Output.MaximumTemperature),
-            ["cm"] = Output.Comments,
+            ["cm"] = FormatHelper.UrlEncode(Output.Comments),
             ["ip"] = FormatHelper.GetValueAsString(Output.PeakEnergyImport),
             ["io"] = FormatHelper.GetValueAsString(Output.OffPeakEnergyImport),
             ["is"] = FormatHelper.GetValueAsString(Output.ShoulderEnergyImport),